Added admin controls

+++ game/socket.py
@@ -21,24 +21,33 @@ def game_handler(io):
         from authorization.models import User
 
         session = io.get_session(sid)
+        user = User.objects.filter(pk=session["id"])
+        user_singular = user.first()
+
+        if not user_singular or user_singular.room:
+            return
+
         room = Room.objects.filter(socket_id=socket_id)
         room_singular = room.first()
 
-        if not room_singular:
+        if not room_singular or room_singular.game_started:
             return
 
-        user = User.objects.filter(pk=session["id"])
-        user_singular = user.first()
-        state = GameState.objects.create(user=user.first(), room=room_singular)
+        state = GameState.objects.filter(user__pk=user_singular.pk).first()
+
+        if state:
+            GameState.reset(state.pk)
+        else:
+            state = GameState.objects.create(user=user_singular, room=room_singular)
+
         user.update(room=room_singular, is_room_admin=False, number_in_room=room_singular.connected + 1)
 
         io.emit(EVENTS_EMIT["JOINED_ROOM"], {
             "id": socket_id,
             "user": user_singular.id,
-            "num_in_room": user_singular.number_in_room,
+            "number_in_room": room_singular.connected,
+            "game_state": state.self_serialize(),
         })
-
-        return socket_id
 
     @io.on(EVENTS_ON["CREATE_ROOM"])
     def create_room(sid, name):
@@ -51,13 +60,20 @@ def game_handler(io):
 
         room = Room.objects.create(name=name, socket_id=socket_id)
         user = User.objects.filter(pk=session["id"])
+        user_singular = user.first()
+
+        if not user_singular:
+            return
+
+        state = GameState.objects.filter(user__pk=user_singular.pk).first()
+
+        if state:
+            GameState.reset(state.pk)
+        else:
+            GameState.objects.create(user=user_singular, room=room)
 
         user.update(room=room, is_room_admin=True, number_in_room=room.connected + 1)
-        state = GameState.objects.create(user=user.first(), room=room)
-
         io.emit(EVENTS_EMIT["CREATED_ROOM"], room.self_serialize())
-
-        return socket_id
 
     @io.on(EVENTS_ON["LEAVE_ROOM"])
     def leave_room(sid):
@@ -67,21 +83,25 @@ def game_handler(io):
         session = io.get_session(sid)
         user = User.objects.filter(pk=session["id"])
         user_singular = user.first()
-        room_id = user_singular.room.pk
+
+        if not user_singular.room or user_singular.is_room_admin:
+            return
+
+        room_id = user_singular.room.socket_id
+        room_pk = user_singular.room.pk
         connected = user_singular.room.connected
 
         if user_singular.room.connected == 1:
-            user_singular.room.delete()
+            Room.objects.filter(pk=room_pk).delete()
 
         GameState.objects.filter(user__pk=user_singular.pk).delete()
         user.update(room=None)
 
         io.emit(EVENTS_EMIT["LEFT_ROOM"], {
             "id": room_id,
+            "user": user_singular.pk,
             "connected": connected - 1,
         })
-
-        return user_singular.pk
 
     @io.on(EVENTS_ON["GET_CURRENT_ROOM"])
     def get_current_room(sid):
@@ -95,3 +115,42 @@ def game_handler(io):
             return
 
         return user.room.self_serialize_detailed(current_user_id=user.pk)
+
+    @io.on(EVENTS_ON["START_GAME"])
+    def start_game(sid):
+        from game.models import Room
+        from authorization.models import User
+
+        session = io.get_session(sid)
+        user = User.objects.filter(pk=session["id"]).first()
+
+        if not user or not user.room or not user.is_room_admin:
+            return
+
+        Room.objects.filter(pk=user.room.pk).update(game_started=True)
+
+        io.emit(EVENTS_EMIT["GAME_STARTED"], {
+            "id": user.room.socket_id,
+            "admin": user.pk,
+        })
+
+    @io.on(EVENTS_ON["CANCEL_GAME"])
+    def cancel_game(sid):
+        from game.models import Room
+        from authorization.models import User
+
+        session = io.get_session(sid)
+        user = User.objects.filter(pk=session["id"]).first()
+
+        if not user or not user.room or not user.is_room_admin:
+            return
+
+        user.room.game_states.all().delete()
+        room_id = user.room.socket_id
+        room_pk = user.room.pk
+        Room.objects.filter(pk=room_pk).delete()
+
+        io.emit(EVENTS_EMIT["GAME_CANCELLED"], {
+            "id": room_id,
+            "admin": user.pk,
+        })
